Reimplemented 1 sweeper per mine, resized window
One sweeper per mine , if a sweeper is killed the mine that it was
targeting is released and another sweeper can target it. If there are
more sweepers than mines, then sweepers are allocated until there are no
more untargeted mines and the remaining sweepers are assigned randomly
to any of the available mines.

Resized window because during tests not all supermines were visible
especially when when testing with 10 supermines.